appsEditServlet

+++ src/main/java/ch/heigvd/amt/mvc/dao/UserApplicationManagerLocal.java
@@ -8,9 +8,9 @@ import java.util.Map;
 
 @Local
 public interface UserApplicationManagerLocal {
-    public UserApplication getApplication(long id);
+    public UserApplication getApplication(int id);
     public ArrayList<UserApplication> getApplicationList(String owner);
     public UserApplication createApplication(String email, String name, String description);
-    public void updateApplication(long id, String name, String description);
-    public void deleteApplication(long id, String owner);
+    public void updateApplication(int id, String name, String description);
+    public void deleteApplication(int id, String owner);
 }
